shell: add process group for unix only
