+++ plugins/builtin/shell/exec_darwin.go
@@ -1,5 +1,10 @@
 package shell
 
+import (
+	"os/exec"
+	"syscall"
+)
+
 const (
 	shellSh         = "/bin/sh"
 	shCommandPrefix = "-c"
@@ -14,4 +19,8 @@ func defaultParams() Params {
 
 func (p *Params) preparedCommand() string {
 	return p.Command
+}
+
+func decorateCommand(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 }
